now create another doc providing instructions on how to implement zustand to use hook use-auth to be used to get user profile and status info on a homepage for example
And also how to integrate with a TanStack to be used in a dashboard.

This is a very commom usage of this auth usage and It will be very helpful to you as a senior developer to easy out this integration process for another users that will join our team and pass this first setup to bolder projects.

Create a new .md file for this task