I have deleted manually the file.
Now we have a loading state for the top right buttons on the header,

But also my main homepage aread is waiting for the loading auth start to render which should not wait for it because it needs to render right away.

The area where it shows:
###
Deliver collaborative
Support
S
u
p
p
o
r
t
Support your customers on Slack, Microsoft Teams, Discord and many more – and move from answering tickets to building genuine relationships.

Your work email
See Nexus in action
Free 14 day trial

Works with
Slack
Teams
Discord
Email
AND MORE
##

Should be rendered ASAP when the user enters the side. It is not a dinamically displayed content.