Great answer. So I will change the name of that folder for /user-profile/ and you do the necessary fixes related to that name change.
Also it is important to document that strucutre and that line of though that you presented when i asked you about the duplication of concers and leave it clear on the approppriade file of documentation, also about the structure of a feature as this one that you did.

I have already changed the name of the folder, so now you do the rest please.