now everything in the backend is working fine. Great job to us.
We just need to adapt some stuff now on the front end now.

We have a Login and Sign Up button on my hero-header

The first time that the page loads, they are clicable but nothing happens... if i go to another page and then get back, it works fine.

What could that be?