it is much better, but the proposed design has the edit together with the display. With a save and cancel at the bottom.
also use the hooks provided and import them to our project. I have already paid for it, so dont be lazy and apply as much consistency to the project as you can.

The filename is profile-view because it is the nomenclature guidelines. It does not mean we cannot edit it the profile info in there.